+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 All notable changes to this project will be documented in this file.
 
+## WIP
+
+- Move the compiler binary to a Symfony command
+
 ## [3.0.0] - 2022-07-28
 [3.0.0]: https://github.com/Smile-SA/gdpr-dump/compare/2.3.3...3.0.0